+++ corehq/apps/oauth_integrations/tests/test_util.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 
 from google.oauth2.credentials import Credentials
 
-from corehq.apps.oauth.models import GoogleApiToken
-from corehq.apps.oauth.utils import get_token, chunkify_data, load_credentials, stringify_credentials
+from corehq.apps.oauth_integrations.models import GoogleApiToken
+from corehq.apps.oauth_integrations.utils import get_token, chunkify_data, load_credentials, stringify_credentials
 
 
 class TestUtils(TestCase):
